app/users.py
import uuid
import os
from typing import Optional
from fastapi import Request, Depends
from fastapi_users import UUIDIDMixin, BaseUserManager, models, FastAPIUsers
from dotenv import load_dotenv
from fastapi_users.authentication import(
    AuthenticationBackend,
    JWTStrategy,
    BearerTransport
)
from fastapi_users.db import SQLAlchemyUserDatabase
from app.db import get_user_db, User

load_dotenv()

Secret = str(os.getenv("SECRET"))

class UserManager(UUIDIDMixin, BaseUserManager[User, uuid.UUID]):
    reset_password_token_secret = Secret
    verification_token_secret = Secret

    async def on_afeter_register(self, user: User, request: Optional[Request] = None):
        print(f"{user.id} has been registered.")
    
    async def on_after_forgot_password(self, user: User, token: str, request: Optional[Request] = None):
        print(f"User {user.id} has forgot their password. Reset token: {token}")

    async def on_after_request_verify(self, user: User, token: str, request: Optional[Request] = None):
        print(f"Verification requested for user {user.id}. Verification token: {token}")

async def get_user_manager(user_db: SQLAlchemyUserDatabase = Depends(get_user_db)):
    yield UserManager(user_db)

bearer_transport = BearerTransport(tokenUrl="auth/jwt/login")

def get_jwt_strategy():
    return JWTStrategy(secret=Secret, lifetime_seconds=3600)

auth_backend = AuthenticationBackend(
    name="jwt",
    transport=bearer_transport,
    get_strategy=get_jwt_strategy
)

fastapi_users = FastAPIUsers[User, uuid.UUID](get_user_manager, [auth_backend])
current_active_user = fastapi_users.current_user(active=True)

    
