pyproject.toml
[project]
name = "fast-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.122.0",
    "fastapi-users[sqlalchemy]>=15.0.1",
    "imagekitio>=4.2.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.52.1",
    "uvicorn[standard]>=0.38.0",
]
